Door_Control_PCB: Add motor driver and connect
Add the TB6612FNG motor driver and connect it to the Photon.  This will
take power from 4 AA batteries in the Lockitron and drive the Lockitron
motor.  The Photon will control direction and speed for the motor.
Since there is only 1 motor on the Lockitron, only 1 of the input/output
pairs on the driver needs to be connected.