Door_Control_PCB: Add headers and diode
Exchange the screw terminals for 0.1" breakaway header holes.  There is
no room for the screw terminals in the Lockitron so wires will need to
be soldered directly to the board.  Some space can be saved by using
0.1" spacing header holes instead of the bigger pitch screw terminals.
Also do a few other minor cleanup tasks like change the size of the wire
input labels.  Also add a 1N4001 diode between VM from AA's and VIN on
Photon to drop the voltage from about 6V to about 5V to ensure it is
safe for the Photon VIN input.